+++ src/decorators/global-event-listener.decorator.ts
@@ -1,5 +1,6 @@
 import {ComponentFactory} from "../helpers/component.factory";
 import {AbstractComponent} from "../classes/abstract.component";
+import {EventHelper} from "../helpers/event-helper";
 
 /**
  * @description
@@ -10,28 +11,7 @@ import {AbstractComponent} from "../classes/abstract.component";
 export function GlobalEventListener(type: string = null) {
 	return function (target: any, propertyKey: string) {
 		ComponentFactory.onComponentClassInitialized(function (object: AbstractComponent) {
-			if (typeof type === 'string') {
-				const events = type.split(' ');
-				for (let i = 0; i < events.length; i++) {
-					applyEvent(document, events[i]);
-				}
-			} else {
-				applyEvent(document);
-			}
-
-			/**
-			 * @description
-			 * Creates a new function that contains the property call with applied scope of the event function
-			 * @param {Element} document
-			 * @param {string} type
-			 */
-			function applyEvent(document: Document, type: string | null = null) {
-				let listener = function (event: Event) {
-					target[propertyKey].apply(object, [this, event]);
-				};
-
-				document.addEventListener(type === null ? propertyKey : type, listener);
-			}
+			EventHelper.attachDecoratorListeners(type, document, target, object, propertyKey);
 		}, target);
 	};
 }