+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.5.0]
+
+### Added
+
+- Simple Facade pattern
+- "Example" Facade for caching (currently just a "VariableCache")
 
 ## [3.4.0]